upgrade images to support arm/arm64
