+++ agent/cita_exporter/.dockerignore
@@ -1,0 +1,1 @@
+.git
