+++ agent/docker-compose-with-build.yml
@@ -1,0 +1,79 @@
+version: "3"
+services:
+
+  citamon_agent_host_exporter:
+    image: prom/node-exporter:v0.18.0
+    hostname: host_exporter
+    container_name: citamon_agent_host_exporter
+    volumes:
+      - /proc:/host/proc
+      - /sys:/host/sys
+      - /:/host/rootfs
+      - /etc/localtime:/etc/localtime
+    command: "--path.procfs='/host/proc' --path.sysfs='/host/sys' --path.rootfs='/host/rootfs'"
+    networks:
+      - citamon-agent-net
+
+  citamon_agent_process_exporter:
+    build:
+      context: ./process_exporter/process-exporter.git
+      dockerfile: Dockerfile
+    image: citamon/agent-process-exporter:0.4.0
+    hostname: process_exporter
+    container_name: citamon_agent_process_exporter
+    volumes:
+      - /proc:/host/proc
+      - ./process_exporter/config:/config
+      - /etc/localtime:/etc/localtime
+    command: "--procfs /host/proc --config.path /config/process_list.yml"
+    networks:
+      - citamon-agent-net
+
+  citamon_agent_rabbitmq_exporter:
+    build: 
+      context: ./rabbitmq_exporter/rabbitmq_exporter.git
+      dockerfile: Dockerfile
+    image: citamon/agent-rabbitmq-exporter:1.0.0-RC
+    hostname: rabbitmq_exporter
+    container_name: citamon_agent_rabbitmq_exporter
+    network_mode: "host"
+    volumes:
+      - /etc/localtime:/etc/localtime
+    environment:
+      - "RABBIT_CAPABILITIES=bert,no_sort"
+
+  citamon_agent_cita_exporter:
+    build:
+      context: ./cita_exporter
+      dockerfile: Dockerfile
+    image: citamon/agent-cita-exporter:20.2.2
+    hostname: ${HOSTNAME}
+    container_name: citamon_agent_cita_exporter_${NODE_IP}_${NODE_PORT}
+    pid: "host"
+    volumes:
+      - ${NODE_DIR}:${NODE_DIR}:ro
+      - ${SOFT_PATH}:${SOFT_PATH}:ro
+      - /etc/localtime:/etc/localtime
+    environment:
+      - NODE_IP_PORT=${NODE_IP}:${NODE_PORT}
+      - NODE_DIR
+    networks:
+      - citamon-agent-net
+
+  citamon_agent_proxy_exporter:
+    image: nginx:1.16.0
+    hostname: proxy_exporter
+    container_name: citamon_agent_proxy_exporter
+    ports:
+      - "1920:80"
+    volumes:
+      - ./nginx.template:/etc/nginx/conf.d/nginx.template
+    environment:
+      - NGINX_HOST=${NODE_IP}
+      - NGINX_PORT=80
+    command: /bin/bash -c "envsubst < /etc/nginx/conf.d/nginx.template > /etc/nginx/conf.d/default.conf && nginx -g 'daemon off;'"
+    networks:
+      - citamon-agent-net
+
+networks:
+  citamon-agent-net:
